requirements.txt
python-dotenv
requests
pandas
tweepy
Pillow
transformers
torch
